+++ python/GafferSceneUI/__init__.py
@@ -185,6 +185,7 @@ from . import ImageToPointsUI
 from . import MeshSplitUI
 from . import FramingConstraintUI
 from . import MeshNormalsUI
+from . import LightToolUI
 
 # then all the PathPreviewWidgets. note that the order
 # of import controls the order of display.